agents/voice_to_text/requirements.txt
openai>=1.0.0
requests>=2.0.0
python-dotenv>=1.0.0
uagents>=0.4.0 